Updated VanillaStyle with new features and bugfixes
Numerous changes have been made to the VanillaStyle codebase. This includes the addition of new features such as support for ToolButtonStyle and LineEditButtonEventFilter. Minor UI tweaks such as changing the background color definition in LightVanillaStyle, optimizing icon and text rendering in the ToolButtonStyle, and adding padding adjustments for icon and checkbox in MenuStyle. Furthermore, a number of miscellaneous bug fixes and performance improvements have been implemented. Changes also include the addition of new icons and resources for UI elements.