Feature/update itemView (#16)
* Refactor styling for MenuItem (#8)

* Refactor styling for MenuItem

* fix: toggle buttion icon

* Feature/QScrollArea background (#11)

* try to change QScroolAreaBackground

* Enhance item view styling and improve table widget setup

* Add splitter support in VanillaStyle

Implemented drawing and sizing for splitters in the VanillaStyle. Updated the UI to include a QSplitter widget and adjusted related settings for consistent style management. Removed commented-out and unused code for better readability and maintenance.

* Update color logic in Helper rendering

Previously, the renderRoundRect function used a hardcoded gray color. Now, it dynamically fetches the button foreground color from the theme, ensuring consistency with the application's theme settings.

* Feature/dot floating (#15)

* Introduce DotFloating widget and custom ItemView background support

- Added a new DotFloating widget with animations.
- Enhanced ItemView background logic with state handling and custom background support.
- Updated dependencies and revised includes for better modularity.
- Added usage of `s_CustomItemViewBackground` property in examples.

---------

Co-authored-by: iveswang <1660583890@qq.com>